# lib/api-zod/src/generated/types/createTeamMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface CreateTeamMemberBody {
  /** @minLength 1 */
  name: string;
  role: Role;
  color?: string;
}