# lib/api-zod/src/generated/types/createVideoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from "./videoStatus";

export interface CreateVideoBody {
  /** @minLength 1 */
  title: string;
  deliveryDate: Date;
  status: VideoStatus;
  client?: string | null;
  platform?: string | null;
  durationSeconds?: number | null;
  notes?: string | null;
  editorId?: number | null;
  captadorId?: number | null;
  roteiristaId?: number | null;
}