# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonContribution } from "./personContribution";
import type { StatusCount } from "./statusCount";
import type { Video } from "./video";

export interface DashboardSummary {
  totalVideos: number;
  deliveredThisWeek: number;
  deliveredThisMonth: number;
  upcomingCount: number;
  byStatus: StatusCount[];
  recentVideos: Video[];
  topContributors: PersonContribution[];
}