# lib/api-zod/src/generated/types/personContribution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMember } from "./teamMember";

export interface PersonContribution {
  member: TeamMember;
  count: number;
}