# lib/api-zod/src/generated/types/statusCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from "./videoStatus";

export interface StatusCount {
  status: VideoStatus;
  count: number;
}