# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface TeamMember {
  id: number;
  name: string;
  role: Role;
  /** Hex color used in UI */
  color: string;
  createdAt: Date;
}