# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMember } from "./teamMember";
import type { VideoStatus } from "./videoStatus";

export interface Video {
  id: number;
  title: string;
  deliveryDate: Date;
  status: VideoStatus;
  client?: string | null;
  /** e.g. Instagram, YouTube, TikTok */
  platform?: string | null;
  durationSeconds?: number | null;
  notes?: string | null;
  editorId?: number | null;
  captadorId?: number | null;
  roteiristaId?: number | null;
  editor?: TeamMember | null;
  captador?: TeamMember | null;
  roteirista?: TeamMember | null;
  createdAt: Date;
}