# lib/api-zod/src/generated/types/videoStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  planejado: "planejado",
  em_producao: "em_producao",
  em_edicao: "em_edicao",
  entregue: "entregue",
  publicado: "publicado",
} as const;