# lib/api-zod/src/generated/types/weeklyReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonContribution } from "./personContribution";
import type { StatusCount } from "./statusCount";
import type { Video } from "./video";
import type { WeeklyReportByDayItem } from "./weeklyReportByDayItem";

export interface WeeklyReport {
  weekStart: Date;
  weekEnd: Date;
  totalVideos: number;
  byStatus: StatusCount[];
  byEditor: PersonContribution[];
  byCaptador: PersonContribution[];
  byRoteirista: PersonContribution[];
  byDay: WeeklyReportByDayItem[];
  videos: Video[];
}